+++ ESFramework/UIKit/UITableView+ESAdditions.h
@@ -12,8 +12,6 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface UITableView (ESAdditions)
 
-- (void)scrollToFirstRow:(BOOL)animated;
-- (void)scrollToLastRow:(BOOL)animated;
 - (void)scrollToFirstResponderAnimated:(BOOL)animated atScrollPosition:(UITableViewScrollPosition)scrollPosition;
 
 - (void)touchRowAtIndexPath:(NSIndexPath *)indexPath animated:(BOOL)animated scrollPosition:(UITableViewScrollPosition)scrollPosition;